requirements.dev.txt
bandit==1.8.3
black==25.1.0
build==1.2.2.post1
dicttoxml==1.7.16
flake8==7.2.0
pipdeptree==2.26.1
pylint==3.3.7
pytest-cov==6.1.1
setuptools==80.3.1
sphinx==8.1.3
furo==2025.12.19
sphinx-multiversion-contrib==0.2.13
tox==4.24.1
twine==6.0.1
wheel==0.46.1
